Fixes floating point output in decorators

+++ header/DrainDecorator.hpp
@@ -11,7 +11,7 @@ class DrainDecorator : public AbilityDecorator {
         BaseCharacter* victim = this->character->attack(charList, attackerIndex);
         if (!(rand() % 3)) {
             this->character->setHealth(character->getHealth() + (victim->getLatestDamage() * (0.1492 + sqrt(0.02512 * this->getLevel() + 0.00573168))/0.01256) / 100);
-            addAbilityMessage(this->character->getName() + " managed to gain " + to_string((victim->getLatestDamage() * (0.1492 + sqrt(0.02512 * this->getLevel() + 0.00573168))/0.01256) / 100) + " from " + victim->getName());
+            addAbilityMessage(this->character->getName() + " managed to gain " + to_string(static_cast<int>((victim->getLatestDamage() * (0.1492 + sqrt(0.02512 * this->getLevel() + 0.00573168))/0.01256) / 100)) + " from " + victim->getName());
             cout << this->currentMessage << endl;
             this->currentMessage = "";
             return victim;
